refactor: extract cli command parsers
