+++ packages/cli/src/index.ts
@@ -61,12 +61,29 @@ import {
   parsePrCommandArgs as parsePrCommandArgsImpl,
   type PrCommandOptions,
 } from "./commands/pr";
+import { parseAuditCommandArgs } from "./commands/audit";
+import {
+  parseFeatureBacklogCommandArgs,
+  parseTestBacklogCommandArgs,
+  type FeatureBacklogCommandOptions,
+  type TestBacklogCommandOptions,
+} from "./commands/backlog";
+import {
+  parseIssueCommandArgs,
+  parseIssueNumber,
+  type IssueCommandOptions,
+  type IssueDraftCommandOptions,
+} from "./commands/issue";
+import {
+  parseReviewCommandArgs,
+  REVIEW_USAGE,
+  type ReviewCommandOptions,
+} from "./commands/review";
 import { listPullRequestsTool } from "./pr-list-tool";
 import { readyPullRequestTool } from "./pr-ready-tool";
 import { parsePrsToolCommandArgs } from "./prs-tool-command";
 import {
   createRepositoryForge,
-  type AuditTarget,
   type CreatedIssueRecord,
   type IssueDetails,
   type IssuePlanComment,
@@ -136,6 +153,13 @@ import { runPrResolveConflictsCommand } from "./workflows/pr-resolve-conflicts/r
 import { runPrFixTestsCommand } from "./workflows/pr-fix-tests/run";
 
 export { parseSetupCommandArgs };
+export { parseAuditCommandArgs } from "./commands/audit";
+export {
+  parseFeatureBacklogCommandArgs,
+  parseTestBacklogCommandArgs,
+} from "./commands/backlog";
+export { parseIssueCommandArgs } from "./commands/issue";
+export { parseReviewCommandArgs } from "./commands/review";
 
 type IssueWorkspace = {
   issueDir: string;
@@ -147,8 +171,6 @@ type IssueWorkspace = {
 };
 
 type IssueWorkspaceMode = "local" | "github-action" | "unattended";
-type IssueRunMode = "interactive" | "unattended";
-type IssuePrepareMode = "local" | "github-action";
 type IssueDraftWorkspace = {
   runDir: string;
   draftFilePath: string;
@@ -160,97 +182,9 @@ type IssueDraftWorkspace = {
   superpowersPlanFilePath: string;
 };
 
-type IssueDraftCommandOptions =
-  | {
-      mode: "caller";
-      draftFilePath?: string;
-      issueSetFilePath?: string;
-      roughIdea?: string;
-      roughIdeaFilePath?: string;
-      contextValues: string[];
-      contextFilePaths: string[];
-      superpowersSpecFilePath?: string;
-      superpowersPlanFilePath?: string;
-    }
-  | {
-      mode: "runtime";
-    };
-
-type IssueCommandOptions =
-  | {
-      action: "run";
-      issueNumber: number;
-      mode: IssueRunMode;
-    }
-  | {
-      action: "batch";
-      issueNumbers: number[];
-      mode: "unattended";
-    }
-  | {
-      action: "prepare";
-      issueNumber: number;
-      mode: IssuePrepareMode;
-    }
-  | {
-      action: "finalize";
-      issueNumber: number;
-      mode: "local";
-    }
-  | {
-      action: "plan";
-      issueNumber: number;
-      mode: "local";
-      refresh: boolean;
-    }
-  | ({
-      action: "draft";
-    } & IssueDraftCommandOptions)
-  | {
-      action: "refine";
-      issueNumber: number;
-    };
-
 type GeneratedIssueResolutionPlan = Awaited<
   ReturnType<typeof generateIssueResolutionPlan>
 >;
-
-type BacklogOutputFormat = "json" | "markdown";
-
-type ReviewOutputFormat = "json" | "markdown";
-
-type ReviewCommandOptions = {
-  base?: string;
-  head?: string;
-  format: ReviewOutputFormat;
-  issueNumber?: number;
-};
-
-export type AuditCommandOptions = {
-  action: "publish";
-  target: AuditTarget;
-  filePath: string;
-  sectionName: string;
-  localRun?: string;
-};
-
-type TestBacklogCommandOptions = {
-  repoRoot: string;
-  format: BacklogOutputFormat;
-  top: number;
-  createIssues: boolean;
-  maxIssues: number;
-  labels: string[];
-};
-
-type FeatureBacklogCommandOptions = {
-  repoRoot: string;
-  format: BacklogOutputFormat;
-  top: number;
-  createIssues: boolean;
-  maxIssues: number;
-  labels: string[];
-};
 
 type IssueRunContext = {
   issueNumber: number;
@@ -401,49 +335,6 @@ type UnattendedIssueRunResult = {
   committed: boolean;
   pullRequest: IssuePullRequestOutcome;
 };
-
-const ISSUE_USAGE = [
-  "Usage:",
-  "  prs issue <number> [--mode <interactive|unattended>]",
-  "  prs issue <number> <number> [...number] [--mode unattended]",
-  "  prs issue batch <number> <number> [...number] [--mode unattended]",
-  "  prs issue draft --draft-file <path> [--rough-idea <text>|--rough-idea-file <path>] [--context <text>] [--context-file <path>] [--superpowers-spec-file <path>] [--superpowers-plan-file <path>]",
-  "  prs issue draft --issue-set-file <path> [--rough-idea <text>|--rough-idea-file <path>] [--context <text>] [--context-file <path>] [--superpowers-spec-file <path>] [--superpowers-plan-file <path>]",
-  "  prs issue draft --runtime",
-  "  prs issue refine <number>",
-  "  prs issue plan <number> [--refresh]",
-  "  prs issue prepare <number> [--mode <local|github-action>]",
-  "  prs issue finalize <number>",
-].join("\n");
-
-const TEST_BACKLOG_USAGE = [
-  "Usage:",
-  "  prs test-backlog [--format <markdown|json>] [--top <count>]",
-  "  prs review tests [--format <markdown|json>] [--top <count>]",
-  "                       [--repo-root <path>] [--create-issues]",
-  "                       [--max-issues <count>] [--label <name>] [--labels <a,b>]",
-].join("\n");
-
-const FEATURE_BACKLOG_USAGE = [
-  "Usage:",
-  "  prs feature-backlog [repo-path] [--format <markdown|json>] [--top <count>]",
-  "  prs review features [repo-path] [--format <markdown|json>] [--top <count>]",
-  "                          [--create-issues] [--max-issues <count>]",
-  "                          [--label <name>] [--labels <a,b>]",
-].join("\n");
-
-const REVIEW_USAGE = [
-  "Usage:",
-  "  prs review [--base <git-ref>] [--head <git-ref>] [--format <markdown|json>]",
-  "                [--issue-number <number>]",
-  "  prs review diff [--base <git-ref>] [--head <git-ref>] [--format <markdown|json>]",
-  "                  [--issue-number <number>]",
-  "  prs review tests [test-backlog options]",
-  "  prs review features [feature-backlog options]",
-].join("\n");
-
-const AUDIT_PUBLISH_USAGE =
-  "Usage: prs audit publish (--issue <number>|--pr <number>) --file <path> --section <name> [--local-run <path>]";
 
 const TOP_LEVEL_HELP = [
   "prs",
@@ -875,433 +766,12 @@ function ensureCleanWorkingTree(repoRoot: string): void {
   }
 }
 
-function parseIssueNumber(rawValue: string | undefined): number {
-  if (!rawValue) {
-    throw new Error(`Missing issue number. ${ISSUE_USAGE}`);
-  }
-
-  if (!/^\d+$/.test(rawValue)) {
-    throw new Error(`Invalid issue number "${rawValue}". ${ISSUE_USAGE}`);
-  }
-
-  const issueNumber = Number.parseInt(rawValue, 10);
-  if (!Number.isSafeInteger(issueNumber) || issueNumber <= 0) {
-    throw new Error(`Invalid issue number "${rawValue}". ${ISSUE_USAGE}`);
-  }
-
-  return issueNumber;
-}
-
-function parseIssueModeOption(rawArgs: string[]): string | undefined {
-  let mode: string | undefined;
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const rawArg = rawArgs[index];
-    if (rawArg === "--mode") {
-      mode = rawArgs[index + 1];
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--mode=")) {
-      mode = rawArg.slice("--mode=".length);
-      continue;
-    }
-
-    throw new Error(`Unknown issue option "${rawArg}". ${ISSUE_USAGE}`);
-  }
-
-  return mode;
-}
-
-function parseIssueRunMode(rawArgs: string[]): IssueRunMode {
-  const mode = parseIssueModeOption(rawArgs);
-  if (mode === undefined) {
-    return "interactive";
-  }
-
-  if (mode !== "interactive" && mode !== "unattended") {
-    throw new Error(
-      `Invalid issue mode "${mode}". Expected "interactive" or "unattended".`
-    );
-  }
-
-  return mode;
-}
-
-function parseIssuePrepareMode(rawArgs: string[]): IssuePrepareMode {
-  const mode = parseIssueModeOption(rawArgs);
-  if (mode === undefined) {
-    return "local";
-  }
-
-  if (mode !== "local" && mode !== "github-action") {
-    throw new Error(
-      `Invalid issue mode "${mode}". Expected "local" or "github-action".`
-    );
-  }
-
-  return mode;
-}
-
-function parseIssuePlanOptions(rawArgs: string[]): { refresh: boolean } {
-  let refresh = false;
-
-  for (const rawArg of rawArgs) {
-    if (rawArg === "--refresh" || rawArg === "--update") {
-      refresh = true;
-      continue;
-    }
-
-    throw new Error(`Unknown issue option "${rawArg}". ${ISSUE_USAGE}`);
-  }
-
-  return { refresh };
-}
-
-function parseIssueBatchArgs(
-  rawArgs: string[],
-  commandLabel = "Multi-issue runs"
-): {
-  issueNumbers: number[];
-  mode: "unattended";
-} {
-  const issueNumbers: number[] = [];
-  let mode: string | undefined;
-
-  for (let index = 0; index < rawArgs.length; index += 1) {
-    const rawArg = rawArgs[index];
-    if (rawArg === "--mode") {
-      mode = rawArgs[index + 1];
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--mode=")) {
-      mode = rawArg.slice("--mode=".length);
-      continue;
-    }
-
-    if (rawArg.startsWith("--")) {
-      throw new Error(`Unknown issue option "${rawArg}". ${ISSUE_USAGE}`);
-    }
-
-    issueNumbers.push(parseIssueNumber(rawArg));
-  }
-
-  if (mode !== undefined && mode !== "unattended") {
-    if (mode === "interactive") {
-      throw new Error(
-        `${commandLabel} only support \`--mode unattended\`. Interactive multi-issue mode is not supported.`
-      );
-    }
-
-    throw new Error(`Invalid issue mode "${mode}". Expected "unattended".`);
-  }
-
-  const uniqueIssueNumbers = [...new Set(issueNumbers)];
-  if (uniqueIssueNumbers.length < 2) {
-    throw new Error(
-      `${commandLabel} require at least two issue numbers. ${ISSUE_USAGE}`
-    );
-  }
-
-  if (uniqueIssueNumbers.length !== issueNumbers.length) {
-    throw new Error(`${commandLabel} do not support duplicate issue numbers.`);
-  }
-
-  return {
-    issueNumbers,
-    mode: "unattended",
-  };
-}
-
-function takeRequiredOptionValue(
-  args: string[],
-  index: number,
-  optionName: string
-): string {
-  const value = args[index + 1];
-  if (!value || value.startsWith("--")) {
-    throw new Error(`Missing value for ${optionName}. ${ISSUE_USAGE}`);
-  }
-
-  return value;
-}
-
-function parseIssueDraftOptions(args: string[]): IssueDraftCommandOptions {
-  let draftFilePath: string | undefined;
-  let issueSetFilePath: string | undefined;
-  let roughIdea: string | undefined;
-  let roughIdeaFilePath: string | undefined;
-  let superpowersSpecFilePath: string | undefined;
-  let superpowersPlanFilePath: string | undefined;
-  const contextValues: string[] = [];
-  const contextFilePaths: string[] = [];
-
-  for (let index = 0; index < args.length; index += 1) {
-    const rawArg = args[index];
-
-    if (rawArg === "--runtime") {
-      if (args.length > 1) {
-        throw new Error(
-          "`prs issue draft --runtime` cannot be combined with caller draft options."
-        );
-      }
-
-      return { mode: "runtime" };
-    }
-
-    if (rawArg === "--from-caller") {
-      continue;
-    }
-
-    if (rawArg === "--draft-file") {
-      draftFilePath = takeRequiredOptionValue(args, index, rawArg);
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--issue-set-file") {
-      issueSetFilePath = takeRequiredOptionValue(args, index, rawArg);
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--rough-idea") {
-      roughIdea = takeRequiredOptionValue(args, index, rawArg);
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--rough-idea-file") {
-      roughIdeaFilePath = takeRequiredOptionValue(args, index, rawArg);
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--context") {
-      contextValues.push(takeRequiredOptionValue(args, index, rawArg));
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--context-file") {
-      contextFilePaths.push(takeRequiredOptionValue(args, index, rawArg));
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--superpowers-spec-file") {
-      superpowersSpecFilePath = takeRequiredOptionValue(args, index, rawArg);
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--superpowers-plan-file") {
-      superpowersPlanFilePath = takeRequiredOptionValue(args, index, rawArg);
-      index += 1;
-      continue;
-    }
-
-    throw new Error(`Unknown issue option "${rawArg}". ${ISSUE_USAGE}`);
-  }
-
-  if (!draftFilePath && !issueSetFilePath) {
-    throw new Error(
-      "`prs issue draft` now ingests a skill-produced draft. Pass --draft-file <path> for one issue, --issue-set-file <path> for linked issues, or --runtime to intentionally open a separate drafting session."
-    );
-  }
-
-  if (draftFilePath && issueSetFilePath) {
-    throw new Error("`prs issue draft` accepts either --draft-file or --issue-set-file, not both.");
-  }
-
-  if (roughIdea && roughIdeaFilePath) {
-    throw new Error("`prs issue draft` accepts either --rough-idea or --rough-idea-file, not both.");
-  }
-
-  return {
-    mode: "caller",
-    draftFilePath,
-    issueSetFilePath,
-    roughIdea,
-    roughIdeaFilePath,
-    contextValues,
-    contextFilePaths,
-    superpowersSpecFilePath,
-    superpowersPlanFilePath,
-  };
-}
-
-export function parseIssueCommandArgs(args: string[]): IssueCommandOptions {
-  const issueArgs = args.slice(1);
-  const subcommand = issueArgs[0];
-
-  if (subcommand === "draft") {
-    return {
-      action: "draft",
-      ...parseIssueDraftOptions(issueArgs.slice(1)),
-    };
-  }
-
-  if (subcommand === "refine") {
-    const optionArgs = issueArgs.slice(2);
-    if (optionArgs.length > 0) {
-      throw new Error(`Unknown issue option "${optionArgs[0]}". ${ISSUE_USAGE}`);
-    }
-
-    return {
-      action: "refine",
-      issueNumber: parseIssueNumber(issueArgs[1]),
-    };
-  }
-
-  if (subcommand === "batch") {
-    const parsed = parseIssueBatchArgs(issueArgs.slice(1), "Batch issue runs");
-    return {
-      action: "batch",
-      issueNumbers: parsed.issueNumbers,
-      mode: parsed.mode,
-    };
-  }
-
-  if (subcommand === "prepare") {
-    return {
-      action: "prepare",
-      issueNumber: parseIssueNumber(issueArgs[1]),
-      mode: parseIssuePrepareMode(issueArgs.slice(2)),
-    };
-  }
-
-  if (subcommand === "finalize") {
-    const optionArgs = issueArgs.slice(2);
-    if (optionArgs.length > 0) {
-      throw new Error(`Unknown issue option "${optionArgs[0]}". ${ISSUE_USAGE}`);
-    }
-
-    return {
-      action: "finalize",
-      issueNumber: parseIssueNumber(issueArgs[1]),
-      mode: "local",
-    };
-  }
-
-  if (subcommand === "plan") {
-    const parsedOptions = parseIssuePlanOptions(issueArgs.slice(2));
-
-    return {
-      action: "plan",
-      issueNumber: parseIssueNumber(issueArgs[1]),
-      mode: "local",
-      refresh: parsedOptions.refresh,
-    };
-  }
-
-  const numericIssueArgs = issueArgs.filter((arg) => /^\d+$/.test(arg));
-  if (numericIssueArgs.length > 1) {
-    const parsed = parseIssueBatchArgs(issueArgs);
-    return {
-      action: "batch",
-      issueNumbers: parsed.issueNumbers,
-      mode: parsed.mode,
-    };
-  }
-
-  return {
-    action: "run",
-    issueNumber: parseIssueNumber(issueArgs[0]),
-    mode: parseIssueRunMode(issueArgs.slice(1)),
-  };
-}
-
 export function parsePrCommandArgs(args: string[]): PrCommandOptions {
   return parsePrCommandArgsImpl(args, parseIssueNumber);
 }
 
 export function parseCodexCommand(args: string[]): CodexCommandOptions {
   return parseCodexCommandArgs(args, parseIssueNumber);
-}
-
-export function parseAuditCommandArgs(args: string[]): AuditCommandOptions {
-  const auditArgs = args[0] === "audit" ? args.slice(1) : args;
-  const subcommand = auditArgs[0];
-
-  if (subcommand !== "publish") {
-    throw new Error(AUDIT_PUBLISH_USAGE);
-  }
-
-  const optionArgs = auditArgs.slice(1);
-  let target: AuditTarget | undefined;
-  let filePath: string | undefined;
-  let sectionName: string | undefined;
-  let localRun: string | undefined;
-
-  for (let index = 0; index < optionArgs.length; index += 1) {
-    const rawArg = optionArgs[index];
-
-    if (rawArg === "--issue" || rawArg === "--pr") {
-      if (target) {
-        throw new Error("`prs audit publish` requires exactly one of --issue or --pr.");
-      }
-
-      target = {
-        type: rawArg === "--issue" ? "issue" : "pull-request",
-        number: parseIssueNumber(optionArgs[index + 1]),
-      };
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--file") {
-      filePath = optionArgs[index + 1];
-      if (!filePath) {
-        throw new Error(AUDIT_PUBLISH_USAGE);
-      }
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--section") {
-      sectionName = optionArgs[index + 1];
-      if (!sectionName) {
-        throw new Error(AUDIT_PUBLISH_USAGE);
-      }
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--local-run") {
-      localRun = optionArgs[index + 1];
-      if (!localRun) {
-        throw new Error(AUDIT_PUBLISH_USAGE);
-      }
-      index += 1;
-      continue;
-    }
-
-    throw new Error(`Unknown audit option "${rawArg}".`);
-  }
-
-  if (!target) {
-    throw new Error("`prs audit publish` requires exactly one of --issue or --pr.");
-  }
-
-  if (!filePath) {
-    throw new Error(AUDIT_PUBLISH_USAGE);
-  }
-
-  if (!sectionName) {
-    throw new Error(AUDIT_PUBLISH_USAGE);
-  }
-
-  return {
-    action: "publish",
-    target,
-    filePath,
-    sectionName,
-    localRun,
-  };
 }
 
 export function parseUpdateCommandArgs(args: string[]): { action: "skills" } {
@@ -1311,366 +781,6 @@ export function parseUpdateCommandArgs(args: string[]): { action: "skills" } {
   }
 
   throw new Error(UPDATE_USAGE);
-}
-
-function parsePositiveInteger(value: string | undefined, flagName: string): number {
-  if (!value || !/^\d+$/.test(value)) {
-    throw new Error(`Invalid value for ${flagName}: "${value ?? ""}". Expected a positive integer.`);
-  }
-
-  const parsedValue = Number.parseInt(value, 10);
-  if (!Number.isSafeInteger(parsedValue) || parsedValue <= 0) {
-    throw new Error(`Invalid value for ${flagName}: "${value}". Expected a positive integer.`);
-  }
-
-  return parsedValue;
-}
-
-export function parseTestBacklogCommandArgs(args: string[]): TestBacklogCommandOptions {
-  const normalizedArgs =
-    args[0] === "review" && args[1] === "tests"
-      ? ["test-backlog", ...args.slice(2)]
-      : args;
-  const optionArgs = normalizedArgs.slice(1);
-  let repoRoot = getDefaultRepoRoot();
-  let format: BacklogOutputFormat = "markdown";
-  let top = 5;
-  let createIssues = false;
-  let maxIssues = 3;
-  const labels = new Set<string>();
-
-  for (let index = 0; index < optionArgs.length; index += 1) {
-    const rawArg = optionArgs[index];
-
-    if (rawArg === "--repo-root") {
-      const rawRepoRoot = optionArgs[index + 1];
-      if (!rawRepoRoot) {
-        throw new Error(`Missing value for --repo-root. ${TEST_BACKLOG_USAGE}`);
-      }
-      repoRoot = resolve(process.cwd(), rawRepoRoot);
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--repo-root=")) {
-      const rawRepoRoot = rawArg.slice("--repo-root=".length);
-      if (!rawRepoRoot) {
-        throw new Error(`Missing value for --repo-root. ${TEST_BACKLOG_USAGE}`);
-      }
-      repoRoot = resolve(process.cwd(), rawRepoRoot);
-      continue;
-    }
-
-    if (rawArg === "--format") {
-      const rawFormat = optionArgs[index + 1];
-      if (rawFormat !== "json" && rawFormat !== "markdown") {
-        throw new Error(`Invalid format "${rawFormat ?? ""}". ${TEST_BACKLOG_USAGE}`);
-      }
-      format = rawFormat;
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--format=")) {
-      const rawFormat = rawArg.slice("--format=".length);
-      if (rawFormat !== "json" && rawFormat !== "markdown") {
-        throw new Error(`Invalid format "${rawFormat}". ${TEST_BACKLOG_USAGE}`);
-      }
-      format = rawFormat;
-      continue;
-    }
-
-    if (rawArg === "--top") {
-      top = parsePositiveInteger(optionArgs[index + 1], "--top");
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--top=")) {
-      top = parsePositiveInteger(rawArg.slice("--top=".length), "--top");
-      continue;
-    }
-
-    if (rawArg === "--create-issues") {
-      createIssues = true;
-      continue;
-    }
-
-    if (rawArg === "--max-issues") {
-      maxIssues = parsePositiveInteger(optionArgs[index + 1], "--max-issues");
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--max-issues=")) {
-      maxIssues = parsePositiveInteger(
-        rawArg.slice("--max-issues=".length),
-        "--max-issues"
-      );
-      continue;
-    }
-
-    if (rawArg === "--label") {
-      const label = optionArgs[index + 1]?.trim();
-      if (!label) {
-        throw new Error(`Missing value for --label. ${TEST_BACKLOG_USAGE}`);
-      }
-      labels.add(label);
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--labels") {
-      const rawLabels = optionArgs[index + 1];
-      if (!rawLabels) {
-        throw new Error(`Missing value for --labels. ${TEST_BACKLOG_USAGE}`);
-      }
-      for (const label of rawLabels.split(",")) {
-        const trimmed = label.trim();
-        if (trimmed) {
-          labels.add(trimmed);
-        }
-      }
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--labels=")) {
-      for (const label of rawArg.slice("--labels=".length).split(",")) {
-        const trimmed = label.trim();
-        if (trimmed) {
-          labels.add(trimmed);
-        }
-      }
-      continue;
-    }
-
-    throw new Error(`Unknown test-backlog option "${rawArg}". ${TEST_BACKLOG_USAGE}`);
-  }
-
-  return {
-    repoRoot,
-    format,
-    top,
-    createIssues,
-    maxIssues: Math.min(maxIssues, top),
-    labels: [...labels],
-  };
-}
-
-export function parseFeatureBacklogCommandArgs(args: string[]): FeatureBacklogCommandOptions {
-  const normalizedArgs =
-    args[0] === "review" && args[1] === "features"
-      ? ["feature-backlog", ...args.slice(2)]
-      : args;
-  const optionArgs = normalizedArgs.slice(1);
-  let repoRoot = getDefaultRepoRoot();
-  let format: BacklogOutputFormat = "markdown";
-  let top = 5;
-  let createIssues = false;
-  let maxIssues = 3;
-  const labels = new Set<string>();
-  let repoPathWasSet = false;
-
-  for (let index = 0; index < optionArgs.length; index += 1) {
-    const rawArg = optionArgs[index];
-
-    if (!rawArg.startsWith("-")) {
-      if (repoPathWasSet) {
-        throw new Error(`Unknown feature-backlog argument "${rawArg}". ${FEATURE_BACKLOG_USAGE}`);
-      }
-
-      repoRoot = resolve(rawArg);
-      repoPathWasSet = true;
-      continue;
-    }
-
-    if (rawArg === "--format") {
-      const rawFormat = optionArgs[index + 1];
-      if (rawFormat !== "json" && rawFormat !== "markdown") {
-        throw new Error(`Invalid format "${rawFormat ?? ""}". ${FEATURE_BACKLOG_USAGE}`);
-      }
-      format = rawFormat;
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--format=")) {
-      const rawFormat = rawArg.slice("--format=".length);
-      if (rawFormat !== "json" && rawFormat !== "markdown") {
-        throw new Error(`Invalid format "${rawFormat}". ${FEATURE_BACKLOG_USAGE}`);
-      }
-      format = rawFormat;
-      continue;
-    }
-
-    if (rawArg === "--top") {
-      top = parsePositiveInteger(optionArgs[index + 1], "--top");
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--top=")) {
-      top = parsePositiveInteger(rawArg.slice("--top=".length), "--top");
-      continue;
-    }
-
-    if (rawArg === "--create-issues") {
-      createIssues = true;
-      continue;
-    }
-
-    if (rawArg === "--max-issues") {
-      maxIssues = parsePositiveInteger(optionArgs[index + 1], "--max-issues");
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--max-issues=")) {
-      maxIssues = parsePositiveInteger(
-        rawArg.slice("--max-issues=".length),
-        "--max-issues"
-      );
-      continue;
-    }
-
-    if (rawArg === "--label") {
-      const label = optionArgs[index + 1]?.trim();
-      if (!label) {
-        throw new Error(`Missing value for --label. ${FEATURE_BACKLOG_USAGE}`);
-      }
-      labels.add(label);
-      index += 1;
-      continue;
-    }
-
-    if (rawArg === "--labels") {
-      const rawLabels = optionArgs[index + 1];
-      if (!rawLabels) {
-        throw new Error(`Missing value for --labels. ${FEATURE_BACKLOG_USAGE}`);
-      }
-      for (const label of rawLabels.split(",")) {
-        const trimmed = label.trim();
-        if (trimmed) {
-          labels.add(trimmed);
-        }
-      }
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--labels=")) {
-      for (const label of rawArg.slice("--labels=".length).split(",")) {
-        const trimmed = label.trim();
-        if (trimmed) {
-          labels.add(trimmed);
-        }
-      }
-      continue;
-    }
-
-    throw new Error(`Unknown feature-backlog option "${rawArg}". ${FEATURE_BACKLOG_USAGE}`);
-  }
-
-  return {
-    repoRoot,
-    format,
-    top,
-    createIssues,
-    maxIssues: Math.min(maxIssues, top),
-    labels: [...labels],
-  };
-}
-
-export function parseReviewCommandArgs(args: string[]): ReviewCommandOptions {
-  const normalizedArgs =
-    args[0] === "review" && args[1] === "diff" ? ["review", ...args.slice(2)] : args;
-  const optionArgs = normalizedArgs.slice(1);
-  let base: string | undefined;
-  let head: string | undefined;
-  let format: ReviewOutputFormat = "markdown";
-  let issueNumber: number | undefined;
-
-  for (let index = 0; index < optionArgs.length; index += 1) {
-    const rawArg = optionArgs[index];
-
-    if (rawArg === "--base") {
-      base = optionArgs[index + 1]?.trim();
-      if (!base) {
-        throw new Error(`Missing value for --base. ${REVIEW_USAGE}`);
-      }
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--base=")) {
-      base = rawArg.slice("--base=".length).trim();
-      if (!base) {
-        throw new Error(`Missing value for --base. ${REVIEW_USAGE}`);
-      }
-      continue;
-    }
-
-    if (rawArg === "--head") {
-      head = optionArgs[index + 1]?.trim();
-      if (!head) {
-        throw new Error(`Missing value for --head. ${REVIEW_USAGE}`);
-      }
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--head=")) {
-      head = rawArg.slice("--head=".length).trim();
-      if (!head) {
-        throw new Error(`Missing value for --head. ${REVIEW_USAGE}`);
-      }
-      continue;
-    }
-
-    if (rawArg === "--format") {
-      const rawFormat = optionArgs[index + 1];
-      if (rawFormat !== "json" && rawFormat !== "markdown") {
-        throw new Error(`Invalid format "${rawFormat ?? ""}". ${REVIEW_USAGE}`);
-      }
-      format = rawFormat;
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--format=")) {
-      const rawFormat = rawArg.slice("--format=".length);
-      if (rawFormat !== "json" && rawFormat !== "markdown") {
-        throw new Error(`Invalid format "${rawFormat}". ${REVIEW_USAGE}`);
-      }
-      format = rawFormat;
-      continue;
-    }
-
-    if (rawArg === "--issue-number") {
-      issueNumber = parseIssueNumber(optionArgs[index + 1]);
-      index += 1;
-      continue;
-    }
-
-    if (rawArg.startsWith("--issue-number=")) {
-      issueNumber = parseIssueNumber(rawArg.slice("--issue-number=".length));
-      continue;
-    }
-
-    throw new Error(`Unknown review option "${rawArg}". ${REVIEW_USAGE}`);
-  }
-
-  if (head && !base) {
-    throw new Error(`--head requires --base. ${REVIEW_USAGE}`);
-  }
-
-  return {
-    base,
-    head,
-    format,
-    issueNumber,
-  };
 }
 
 function resolveLaunchStageNoticeId(args: string[]): LaunchStageNoticeId | undefined {
